room management done

+++ DBForm.h
@@ -74,7 +74,7 @@ protected:
 	bool m_is_connect_;        /* 是否连接数据库，true表示已连上 */
     std::string m_query_sql_;  /* SQL查询语句 */
 	SQLINTEGER m_sql_pro_ret;     /* the return of procedure in database*/
-	int m_pro_ret;              /* the return of procedure*/
+	short m_pro_ret;              /* the return of procedure*/
 private:
     char m_datetime_[20];         /* 日期时间 */
 	SQLINTEGER m_sql_datetime_;  /**/
